image/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1

opencv-python~=4.10.0.84
numpy~=2.1.3
pillow~=10.3.0
ultralytics~=8.3.28
scikit-learn~=1.5.2

seaborn~=0.13.2

flask~=3.0.3
pytorch-lightning~=2.4.0